+++ src/scripts/src/InstanceScripts/Instance_Mauradon.cpp
@@ -44,7 +44,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -53,7 +53,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -62,7 +62,7 @@ public:
 		     spells[2].cooldown = 10;
 		     spells[2].targettype = TARGET_ATTACKING;
 		     spells[2].instant = true;
-		     spells[2].perctrigger = (float)RandomFloat(20.0f);
+		     spells[2].perctrigger = RandomFloat(20.0f);
 		     spells[2].attackstoptimer = 3000;
 		     m_spellcheck[2] = true;
 
@@ -100,7 +100,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -174,7 +174,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -183,7 +183,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -221,7 +221,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -295,7 +295,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -304,7 +304,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -342,7 +342,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -416,7 +416,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -425,7 +425,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -434,7 +434,7 @@ public:
 		     spells[2].cooldown = 10;
 		     spells[2].targettype = TARGET_ATTACKING;
 		     spells[2].instant = true;
-		     spells[2].perctrigger = (float)RandomFloat(20.0f);
+		     spells[2].perctrigger = RandomFloat(20.0f);
 		     spells[2].attackstoptimer = 3000;
 		     m_spellcheck[2] = true;
 
@@ -443,7 +443,7 @@ public:
 		     spells[3].cooldown = 10;
 		     spells[3].targettype = TARGET_ATTACKING;
 		     spells[3].instant = true;
-		     spells[3].perctrigger = (float)RandomFloat(20.0f);
+		     spells[3].perctrigger = RandomFloat(20.0f);
 		     spells[3].attackstoptimer = 3000;
 		     m_spellcheck[3] = true;
 
@@ -481,7 +481,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -555,7 +555,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -564,7 +564,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -573,7 +573,7 @@ public:
 		     spells[2].cooldown = 10;
 		     spells[2].targettype = TARGET_ATTACKING;
 		     spells[2].instant = true;
-		     spells[2].perctrigger = (float)RandomFloat(20.0f);
+		     spells[2].perctrigger = RandomFloat(20.0f);
 		     spells[2].attackstoptimer = 3000;
 		     m_spellcheck[2] = true;
 
@@ -611,7 +611,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -685,7 +685,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -694,7 +694,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -732,7 +732,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -806,7 +806,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -815,7 +815,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -853,7 +853,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
